+++ gwbackupy/__main__.py
@@ -1,0 +1,4 @@
+from gwbackupy import gwbackupy
+
+if __name__ == "__main__":
+    gwbackupy.cli_startup()
