+++ gwbackupy/filters/filter_interface.py
