+++ gwbackupy/global_properties.py
