+++ scripts/gwbackupy
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
 
-from gwb import gwbackupy
+from gwbackupy import gwbackupy
 
 gwbackupy.cli_startup()
