+++ setup.py
@@ -1,5 +1,5 @@
 from setuptools import setup
-import gwb.global_properties as global_properties
+import gwbackupy.global_properties as global_properties
 
 with open("README.md", "r") as fh:
     long_description = fh.read()
@@ -12,7 +12,7 @@ with open("requirements.txt", "r") as fh:
 setup(
     name="gwbackupy",
     version=global_properties.version,
-    packages=["gwb", "gwb.storage", "gwb.filters"],
+    packages=["gwbackupy", "gwbackupy.storage", "gwbackupy.filters"],
     url="https://github.com/smartondev/gwbackupy",
     license='BSD 3-Clause "New" or "Revised" License',
     author="Márton Somogyi",
